IKASAN-1314: Remove ikasan-exclusion-service from pom, move IsExclusionServiceAware to exclusion spec project
