+++ ikasaneip/component/endpoint/jms-spring/src/main/java/org/springframework/jms/listener/IkasanMessageListenerContainer.java
@@ -43,7 +43,7 @@ package org.springframework.jms.listener;
 import org.ikasan.component.endpoint.jms.consumer.MessageProvider;
 import org.ikasan.component.endpoint.jms.spring.consumer.IkasanListMessage;
 import org.ikasan.component.endpoint.jms.spring.consumer.SpringMessageConsumerConfiguration;
-import org.ikasan.exclusion.service.IsExclusionServiceAware;
+import org.ikasan.spec.exclusion.IsExclusionServiceAware;
 import org.ikasan.spec.configuration.Configured;
 import org.ikasan.spec.exclusion.ExclusionService;
 import org.springframework.jms.util.JndiUtils;
